backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.11.0
httpx>=0.26.0,<0.29.0
python-multipart>=0.0.6
openai>=1.0.0
python-dotenv>=1.0.0
supabase>=2.9.0
pymongo>=4.6.0
